Adjusted interface of libxsmm_shuffle.

+++ src/libxsmm_math.c
@@ -630,15 +630,15 @@ LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_matdiff_clear)(libxsmm_matdiff_info* in
 }
 
 
-LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_shuffle)(int* /*coprime*/, const int* /*n*/);
-LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_shuffle)(int* coprime, const int* n)
+LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_shuffle)(long long* /*coprime*/, const int* /*n*/);
+LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_shuffle)(long long* coprime, const int* n)
 {
 #if !defined(NDEBUG)
   static int error_once = 0;
   if (NULL != coprime && NULL != n)
 #endif
   {
-    *coprime = (libxsmm_shuffle(*n) & 0x7FFFFFFF);
+    *coprime = (long long)(libxsmm_shuffle(*n) & 0x7FFFFFFF);
   }
 #if !defined(NDEBUG)
   else if (0 != libxsmm_verbosity /* library code is expected to be mute */
@@ -659,7 +659,7 @@ LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_hash)(int* hash, const void* data, cons
   if (NULL != hash && NULL != data && NULL != size && NULL != seed)
 #endif
   {
-    *hash = (libxsmm_hash(data, *size, *seed) & 0x7FFFFFFF);
+    *hash = (int)(libxsmm_hash(data, *size, *seed) & 0x7FFFFFFF);
   }
 #if !defined(NDEBUG)
   else if (0 != libxsmm_verbosity /* library code is expected to be mute */
@@ -680,7 +680,7 @@ LIBXSMM_API void LIBXSMM_FSYMBOL(libxsmm_hash2)(long long* hash, const void* dat
   if (NULL != hash && NULL != data && NULL != size && NULL != seed)
 #endif
   {
-    *hash = (libxsmm_hash(data, *size, *seed) & 0x7FFFFFFFFFFFFFFF);
+    *hash = (long long)(libxsmm_hash(data, *size, *seed) & 0x7FFFFFFFFFFFFFFF);
   }
 #if !defined(NDEBUG)
   else if (0 != libxsmm_verbosity /* library code is expected to be mute */
